Multiplayer

+++ core/src/main/java/com/diamantino/voxelcraft/common/world/WorldSettings.java
@@ -1,5 +1,5 @@
 package com.diamantino.voxelcraft.common.world;
 
-public record WorldSettings(int seed) {
+public record WorldSettings(long seed) {
 
 }
